% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCorrelation.R
\name{SampleCorrelation}
\alias{SampleCorrelation}
\title{Sample Correlation Function}
\usage{
SampleCorrelation(X)
}
\arguments{
\item{X}{an input matrix of size (n x p)}
}
\value{
a matrix of correlation
}
\description{
Takes a matrix and returns a correlation matrix
}
\examples{
X = matrix(c(15,40,37,31,19,47,45,74,16), nrow = 3, ncol = 3, byrow = TRUE): SampleCorrelation(X)
}
