% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleCovariance.R
\name{SampleCovariance}
\alias{SampleCovariance}
\title{Sample Covariance Function}
\usage{
SampleCovariance(X)
}
\arguments{
\item{X}{an input matrix of size (n x p)}
}
\value{
a matrix of covariance
}
\description{
Takes a matrix and returns a covariance matrix
}
\examples{
X = matrix(c(15,40,37,31,19,47,45,74,16), nrow = 3, ncol = 3, byrow = TRUE): SampleCovariance(X)
}
