% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMean.R
\name{SampleMean}
\alias{SampleMean}
\title{Sample Mean Function}
\usage{
SampleMean(X)
}
\arguments{
\item{X}{an input matrix of size (n x p)}
}
\value{
a vector of means
}
\description{
Takes a matrix and returns a vector of the column means
}
\examples{
X = matrix(c(15,40,37,31,19,47,45,74,16), nrow = 3, ncol = 3, byrow = TRUE): SampleMean(X)
}
